requirements.txt
certifi
charset-normalizer
dill
idna
numpy
psycopg2-binary
python-decouple
python-dotenv
requests
tqdm
urllib3
uuid6
schedule